+++ src/components/NavBar/NavBar.js
@@ -1,0 +1,32 @@
+import React from 'react';
+
+import AppBar from '@mui/material/AppBar';
+import Box from '@mui/material/Box';
+import Button from '@mui/material/Button';
+import Grid from '@mui/material/Grid';
+import Divider from '@mui/material/Divider';
+import Toolbar from '@mui/material/Toolbar';
+import Stack from '@mui/material/Stack';
+import Link from '@mui/material/Link';
+
+import makeStyles from '@mui/styles/makeStyles';
+
+import MenuIcon from '@mui/icons-material/Menu';
+
+const sections = [
+  { to: '/', text: 'Home' },
+  { to: '/class-schedule', text: 'Class Schedule' },
+  { to: '/membership', text: 'Membership' },
+  { to: '/instructors', text: 'Instructors' },
+  { to: '/contact-info', text: 'Contact Info' },
+];
+
+function NavBar() {
+  return (
+    <nav>
+      NavBar
+    </nav>
+  );
+}
+
+export default NavBar;
